simplify slither printer logic
