--- fuzzing/integrations/slither/testdata/contracts/slither_constants.sol
@@ -1,15 +1,0 @@
-contract SlitherConstants {
-    uint x;
-
-    function f(uint i) public payable {
-        x = i;
-    }
-
-    function echidna_uint() public view returns (bool) {
-        return x != uint(-2);
-    }
-
-    function echidna_ether() public view returns (bool) {
-        return x != 2 * 2 ether;
-    }
-}
